refactoring: moves Interfaces EvaluateWithRemoteRepos, etc. to own file
