+++ src/commands/repos/WriteRepos.test.ts
@@ -1,11 +1,7 @@
-import {IReposDescriptor, IRepoStatus} from './models';
-import * as fs from 'fs';
-import * as path from 'path';
-import {withTempDirectory} from '../../test/helpers/directories';
-import * as child_process from 'child_process';
+import {IRepoStatus} from './models';
 import {WriteRepos} from './WriteRepos';
-import {enforceNewline} from '../../util';
 import {ICommandParameters} from '../models';
+import {basicTestSetupData, testWith} from '../../test/helpers/remoteRepositories';
 
 describe('writing the parent repos json', () => {
 
@@ -114,247 +110,3 @@ describe('writing the parent repos json', () => {
             .evaluate(testUnFreeze, assertRaw);
     });
 });
-
-function testWith(testSetupData: ITestSetupData): ITestRun {
-    return new EvaluateWithRemoteRepos(testSetupData);
-}
-
-interface ITestRun {
-    withBranchUnderTest(branchUnderTest: string): ITestRun;
-
-    withDebug(debug: boolean): ITestRun;
-
-    evaluate(testCase: (rootDir: string) => Promise<void>, assertion: (parentRepos: string) => Promise<void>): Promise<void>;
-}
-
-export interface ITestSetupData {
-    rootRepo: IRepoReleaseTestSetupData;
-
-    [repoName: string]: IRepoReleaseTestSetupData;
-}
-
-export interface IBranchReleaseTestSetupData {
-    branchName: string;
-    releases?: string[];
-}
-
-export interface IRepoReleaseTestSetupData {
-    repoName: string;
-    releaseBranches: IBranchReleaseTestSetupData[];
-}
-
-interface ILocalRepoData {
-    name: string;
-    url: string;
-}
-
-export const ROOT_REPO = 'rootRepo';
-
-export const basicTestSetupData: ITestSetupData = {
-    rootRepo: {
-        repoName: 'foo',
-        releaseBranches: [{branchName: 'release/22.2', releases: ['version/22.2.0']}]
-    },
-    main: {
-        repoName: 'main',
-        releaseBranches: [{branchName: 'release/22.2', releases: ['version/22.2.0']}]
-    },
-    test_1: {
-        repoName: 'test-1',
-        releaseBranches: [{branchName: 'release/22.2', releases: ['version/22.2.0']}]
-    },
-    test_2: {
-        repoName: 'test-2', releaseBranches: [{branchName: 'release/22.2', releases: ['version/22.2.0']}]
-    }
-};
-
-export class EvaluateWithRemoteRepos implements ITestRun {
-
-    private debug: boolean = false;
-    private readonly testSetupData: ITestSetupData;
-    private branchUnderTest: string;
-    private readonly branchesToCheckout: string[] = [];
-
-    constructor(testSetupData: ITestSetupData) {
-        this.testSetupData = testSetupData;
-    }
-
-    private static writeParentRepos(rootDir: string, newParentRepos: IReposDescriptor): void {
-        const filtered: IReposDescriptor = {};
-        Object.keys(newParentRepos).forEach((name) => {
-            if (name !== ROOT_REPO) {
-                filtered[name] = newParentRepos[name];
-            }
-        });
-        const newParentReposContent = enforceNewline(JSON.stringify(filtered, null, 2));
-        const parentRepos = path.join(rootDir, 'parent-repos.json');
-        fs.writeFileSync(parentRepos, newParentReposContent, 'utf8');
-    }
-
-    public withBranchUnderTest(branchUnderTest: string): ITestRun {
-        this.branchUnderTest = branchUnderTest;
-        return this;
-    }
-
-    public withBranchesToCheckout(branchesToCheckout: string[]): ITestRun {
-        this.branchesToCheckout.push(...branchesToCheckout);
-        return this;
-    }
-
-    public withDebug(debug: boolean): ITestRun {
-        this.debug = debug;
-        return this;
-    }
-
-    public evaluate(testCase: (rootDir: string) => Promise<void>, assertion: (parentRepos: string) => Promise<void>): Promise<void> {
-        return withTempDirectory('freeze-parent-repos', this.createRemoteRepos, testCase, assertion).then(
-            () => Promise.resolve(),
-            (e) => {
-                console.log('Failed assertion or error while evaluating a test!', e);
-                // fail in case of an exception
-                expect(e).toBeUndefined();
-                Promise.reject(e);
-            });
-    }
-
-    private createRemoteRepos = async (dir: string, testCase: (rootDir: string) => Promise<void>, assertion: (parentRepos: string) => Promise<void>) => {
-        const remotePath = path.join(dir, 'remote');
-        fs.mkdirSync(remotePath);
-        const remoteRepos = this.createRepositories(Object.keys(this.testSetupData), remotePath);
-
-        if (!Array.isArray(remoteRepos)) {
-            throw new Error('Failed to create test setup!');
-        }
-
-        const workingDir = path.join(dir, 'working');
-        fs.mkdirSync(workingDir);
-
-        // set up remotes
-        for (const remote of remoteRepos) {
-            this.testSetupData[remote.name].releaseBranches.forEach((releaseBranch) => {
-                this.debugLog(`setting up remote repository ${remote.name}`);
-                this.branchOff(remote, releaseBranch.branchName);
-                releaseBranch.releases.forEach((release) => this.createRelease(remote, release));
-            });
-        }
-
-        // set up working copy
-        const remoteRootRepo = remoteRepos.find((repo) => repo.name === this.testSetupData.rootRepo.repoName);
-        if (!remoteRootRepo) {
-            // tslint:disable-next-line:max-line-length
-            throw new Error(`Remote repo name in test setup data [${this.testSetupData.rootRepo.repoName}] does not match any name of created repo names [${Object.values(remoteRepos).map((r) => r.name).join(', ')}]!`);
-        }
-        const rootDir = remoteRootRepo.url;
-        const repos = this.initParentRepos(rootDir, remoteRepos, this.branchUnderTest);
-
-        for (const remote of remoteRepos) {
-            this.cloneRepo(workingDir, repos[remote.name].url);
-            this.checkoutBranch(path.join(workingDir, remote.name), this.branchUnderTest);
-        }
-
-        // set up rootRepo
-        for (const branch of this.branchesToCheckout) {
-            this.checkoutBranch(path.join(workingDir, ROOT_REPO), branch);
-        }
-        this.checkoutBranch(path.join(workingDir, ROOT_REPO), this.branchUnderTest);
-
-        if (this.debug) {
-            this.debugLog(`initial parent repos`, fs.readFileSync(path.join(rootDir, 'parent-repos.json')));
-        }
-
-        try {
-            await testCase(rootDir);
-        } catch (e) {
-            this.debugLog('Unexpected error during test evaluation!', e);
-            throw e;
-        }
-
-        const result = fs.readFileSync(path.join(rootDir, 'parent-repos.json'));
-        await assertion(result.toString());
-    }
-
-    private createRelease(localRepoData: ILocalRepoData, tag: string): void {
-        this.commitSomeChanges(localRepoData, 'test_1.txt');
-        this.tagRepo(localRepoData.url, tag);
-    }
-
-    private initParentRepos(rootDir: string, repoData: ILocalRepoData[], branchName: string): IReposDescriptor {
-        const repos: IReposDescriptor = {};
-        repoData.forEach((localRepoData) => {
-            repos[localRepoData.name] = {
-                url: `file://${localRepoData.url}/.git`,
-                branch: branchName
-            };
-        });
-        EvaluateWithRemoteRepos.writeParentRepos(rootDir, repos);
-        this.updateRepo(rootDir);
-        return repos;
-    }
-
-    private branchOff(localRepoData: ILocalRepoData, releaseBranch: string): void {
-        this.commitSomeChanges(localRepoData);
-        this.checkoutBranch(localRepoData.url, releaseBranch);
-    }
-
-    private commitSomeChanges(localRepoData: ILocalRepoData, fileName: string = 'test.txt'): void {
-        fs.writeFileSync(`${localRepoData.url}/${fileName}`, 'test', 'utf8');
-        this.updateRepo(localRepoData.url);
-    }
-
-    // tslint:disable-next-line:no-any
-    private debugLog(message?: any, ...args: any[]): void {
-        if (this.debug) {
-            console.log(message, ...args);
-        }
-    }
-
-    private execSync(pathToRepo: string, command: string): void {
-        this.debugLog(child_process.execSync(
-            command,
-            {
-                cwd: pathToRepo,
-                shell: 'bash'
-            }
-        ).toString());
-    }
-
-    private createRepositories(repos: string[], rootDir: string): ILocalRepoData[] {
-        const command = `git init && git commit --no-gpg-sign --allow-empty -m "empty" `;
-        const data: ILocalRepoData[] = [];
-        for (const repo of repos) {
-            const pathToRepo = path.join(rootDir, repo);
-            fs.mkdirSync(pathToRepo);
-            this.debugLog(`init repo ${repo} in ${pathToRepo}`);
-            this.execSync(pathToRepo, command);
-            data.push({
-                          name: repo,
-                          url: pathToRepo
-                      });
-        }
-        return data;
-    }
-
-    private checkoutBranch(pathToRepo: string, branch: string): void {
-        const command = `git checkout -B "${branch}"`;
-        this.debugLog(`checkout branch in ${pathToRepo}: ${command}`);
-        this.execSync(pathToRepo, command);
-    }
-
-    private updateRepo(pathToRepo: string): void {
-        const command = `git status && git add '.' && git commit --no-gpg-sign --allow-empty -m "empty"`;
-        this.debugLog(`update repo in ${pathToRepo}`);
-        this.execSync(pathToRepo, command);
-    }
-
-    private tagRepo(pathToRepo: string, version: string): void {
-        const command = `git tag -a "${version}" -m "empty"`;
-        this.debugLog(`tag repo in ${pathToRepo} to ${version}`);
-        this.execSync(pathToRepo, command);
-    }
-
-    private cloneRepo(pathToRepo: string, repoUrl: string): void {
-        const command = `git clone ${repoUrl}`;
-        this.debugLog(`cloning repo ${repoUrl} in ${pathToRepo}`);
-        this.execSync(pathToRepo, command);
-    }
-}
